+++ src/apis/assetFilesAPI.js
@@ -1,4 +1,4 @@
-import { GET } from 'utils/http';
+import { GET } from 'apis/http';
 
 export const fetchFile = (fileDetails) => {
   let responseType = 'json';
