+++ src/utils/routes/files/index.js
@@ -1,0 +1,1 @@
+export * from './fileUtils';
