+++ webpack.config.js
@@ -22,7 +22,7 @@ module.exports = {
     extensions: ['.js', '.json', '.jsx', '.css', '.scss'],
     alias: {
       actions: path.resolve(__dirname, 'src/actions'),
-      api: path.resolve(__dirname, 'src/api'),
+      apis: path.resolve(__dirname, 'src/apis'),
       assets: path.resolve(__dirname, 'src/assets'),
       components: path.resolve(__dirname, 'src/components'),
       config: path.resolve(__dirname, 'src/config'),
